requirements.txt
fxpmath
matplotlib
numpy
PyYAML
rawpy
scipy
tqdm